feat: add new hyulights fixtures to the db
